Actualizar ObtenerPreciosAction.php, Config.php y 14 más archivos...

+++ src/Migrations/2019_07_18_186236_create_activos_mercados_table.php
@@ -8,8 +8,9 @@ class CreateActivosMercadosTable extends Migration
 {
     public function up()
     {
-        Schema::create('activo_mercado', function (Blueprint $table) {
-            $table->bigIncrements('id');
+        Schema::create('activo_mercado', function (Blueprint $table) 
+        {
+            $table->increments('id');
             $table->integer('mercado_id')->unsigned()->refers('id')->on('mercados');
             $table->integer('activo_id')->unsigned()->refers('id')->on('activos');
             $table->timestamps();
